Extract stats into a module
